Add RARCH_DISPLAY_CTL_GET_NEXT_VIDEO_OUT

+++ gfx/video_context_driver.c
@@ -243,7 +243,7 @@ bool gfx_ctx_check_window(void *data, bool *quit, bool *resize,
    if (!data)
       return false;
    
-   ctx->check_window(data, quit, resize, width, height, *frame_count);
+   ctx->check_window(data, quit, resize, width, height, (unsigned int)*frame_count);
 
    return true;
 }
